Aenderungen all around

+++ src/model/Artikel.java
@@ -107,7 +107,7 @@ public class Artikel {
         }
     }
 
-    public void rabattieren(int rabatt) {
+    public void rabattieren(float rabatt) {
         this.rabatt = rabatt;
     }
